# docs/swagger.yaml
basePath: /
definitions:
...
  press_release.CreatePressReleaseRequest:
    properties:
      authorId:
        type: integer
      categoryId:
        type: integer
      content:
        minLength: 100
        type: string
      excerpt:
        maxLength: 500
        minLength: 50
        type: string
      location:
        type: string
      metadata:
        $ref: '#/definitions/press_release.PressReleaseMetadata'
      publishDate:
        type: string
      status:
        allOf:
        - $ref: '#/definitions/press_release.PressReleaseStatus'
        enum:
        - draft
        - review
        - published
      tags:
        type: string
      title:
        maxLength: 200
        minLength: 10
        type: string
    required:
    - authorId
    - categoryId
    - content
    - excerpt
    - publishDate
    - status
    - title
    type: object
...
  press_release.PressReleaseListResponse:
    properties:
      limit:
        type: integer
      page:
        type: integer
      pressReleases:
        items:
          $ref: '#/definitions/press_release.PressRelease'
        type: array
      total:
        type: integer
      totalPages:
        type: integer
    type: object
...
  press_release.PressReleaseResponse:
    properties:
      pressRelease:
        $ref: '#/definitions/press_release.PressRelease'
    type: object
...
  press_release.UpdatePressReleaseRequest:
    properties:
      authorId:
        type: integer
      categoryId:
        type: integer
      content:
        minLength: 100
        type: string
      excerpt:
        maxLength: 500
        minLength: 50
        type: string
      location:
        type: string
      metadata:
        $ref: '#/definitions/press_release.PressReleaseMetadata'
      publishDate:
        type: string
      status:
        allOf:
        - $ref: '#/definitions/press_release.PressReleaseStatus'
        enum:
        - draft
        - review
        - published
      tags:
        type: string
      title:
        maxLength: 200
        minLength: 10
        type: string
    type: object
...
  response.Meta:
    properties:
      cursor:
        type: string
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  response.Response:
    properties:
      data: {}
      error:
        type: string
      meta:
        $ref: '#/definitions/response.Meta'
      success:
        type: boolean
    type: object
...
host: 192.168.1.2:8081
info:
  contact:
    email: support@example.com
    name: API Support
  description: RESTful API for healthcare market research reports and categories with
    caching and pagination support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://example.com/terms/
  title: Healthcare Market Research API
  version: "1.0"
paths:
...
  /api/v1/press-releases:
    get:
      consumes:
      - application/json
      description: Get a paginated list of press releases with optional filtering
      parameters:
      - description: 'Filter by status: draft, review, published'
        in: query
        name: status
        type: string
      - description: Filter by category ID
        in: query
        name: categoryId
        type: integer
      - description: Filter by tags (comma-separated)
        in: query
        name: tags
        type: string
      - description: Filter by author ID
        in: query
        name: authorId
        type: integer
      - description: Filter by location
        in: query
        name: location
        type: string
      - description: Search in title, excerpt, content
        in: query
        name: search
        type: string
      - description: 'Page number (default: 1, min: 1)'
        in: query
        name: page
        type: integer
      - description: 'Items per page (default: 20, max: 100)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of press releases with pagination
          schema:
            $ref: '#/definitions/press_release.PressReleaseListResponse'
        "500":
          description: Internal server error
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                error:
                  type: string
              type: object
      summary: Get all press releases
      tags:
      - PressReleases
    post:
      consumes:
      - application/json
      description: Create a new press release
      parameters:
      - description: Press release data
        in: body
        name: pressRelease
        required: true
        schema:
          $ref: '#/definitions/press_release.CreatePressReleaseRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Press release created successfully
          schema:
            $ref: '#/definitions/press_release.PressReleaseResponse'
        "400":
          description: Bad request - invalid input or validation error
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Internal server error
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                error:
                  type: string
              type: object
      summary: Create press release
      tags:
      - PressReleases
  /api/v1/press-releases/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a press release by ID
      parameters:
      - description: Press release ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Press release deleted successfully
        "400":
          description: Bad request - invalid ID
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                error:
                  type: string
              type: object
        "404":
          description: Press release not found
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Internal server error
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                error:
                  type: string
              type: object
      summary: Delete press release
      tags:
      - PressReleases
    get:
      consumes:
      - application/json
      description: Get a single press release by ID
      parameters:
      - description: Press release ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Press release details
          schema:
            $ref: '#/definitions/press_release.PressReleaseResponse'
        "400":
          description: Bad request - invalid ID
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                error:
                  type: string
              type: object
        "404":
          description: Press release not found
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Internal server error
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                error:
                  type: string
              type: object
      summary: Get press release by ID
      tags:
      - PressReleases
    put:
      consumes:
      - application/json
      description: Update an existing press release
      parameters:
      - description: Press release ID
        in: path
        name: id
        required: true
        type: integer
      - description: Press release update data
        in: body
        name: pressRelease
        required: true
        schema:
          $ref: '#/definitions/press_release.UpdatePressReleaseRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Press release updated successfully
          schema:
            $ref: '#/definitions/press_release.PressReleaseResponse'
        "400":
          description: Bad request - invalid input or validation error
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                error:
                  type: string
              type: object
        "404":
          description: Press release not found
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Internal server error
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                error:
                  type: string
              type: object
      summary: Update press release
      tags:
      - PressReleases
  /api/v1/press-releases/{id}/publish:
    patch:
      consumes:
      - application/json
      description: Change press release status to published
      parameters:
      - description: Press release ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Press release published successfully
          schema:
            $ref: '#/definitions/press_release.PressReleaseResponse'
        "400":
          description: Bad request - invalid ID or press release cannot be published
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                error:
                  type: string
              type: object
        "404":
          description: Press release not found
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Internal server error
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                error:
                  type: string
              type: object
      summary: Publish press release
      tags:
      - PressReleases
  /api/v1/press-releases/{id}/submit-review:
    patch:
      consumes:
      - application/json
      description: Change press release status from draft to review
      parameters:
      - description: Press release ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Press release submitted for review successfully
          schema:
            $ref: '#/definitions/press_release.PressReleaseResponse'
        "400":
          description: Bad request - invalid ID or press release cannot be submitted
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                error:
                  type: string
              type: object
        "404":
          description: Press release not found
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Internal server error
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                error:
                  type: string
              type: object
      summary: Submit press release for review
      tags:
      - PressReleases
  /api/v1/press-releases/{id}/unpublish:
    patch:
      consumes:
      - application/json
      description: Change press release status from published back to draft
      parameters:
      - description: Press release ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Press release unpublished successfully
          schema:
            $ref: '#/definitions/press_release.PressReleaseResponse'
        "400":
          description: Bad request - invalid ID or press release is not published
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                error:
                  type: string
              type: object
        "404":
          description: Press release not found
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Internal server error
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                error:
                  type: string
              type: object
      summary: Unpublish press release
      tags:
      - PressReleases
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...